AzureFoundryAgent/requirements.txt
azure-ai-projects>=1.0.0b4
azure-identity>=1.15.0
python-dotenv>=1.0.0